+++ native/opencv-wrapper.cc
@@ -296,13 +296,6 @@ cv_imencode(const char* const ext, const cv::Mat* const image, const int* const 
 // =============================================================================
 //   Highgui: high-level GUI
 // =============================================================================
-void cv_named_window(const char* const winname, int flags) {
-    cv::namedWindow(winname, flags);
-}
-
-void cv_destroy_window(const char* const winname) {
-    cv::destroyWindow(winname);
-}
 
 void cv_imshow(const char* const winname, cv::Mat* mat) {
     if (mat != NULL) {
@@ -312,10 +305,6 @@ void cv_imshow(const char* const winname, cv::Mat* mat) {
 
 int cv_wait_key(int delay) {
     return cv::waitKey(delay);
-}
-
-void cv_set_mouse_callback(const char* const winname, cv::MouseCallback on_mouse, void* userdata) {
-    cv::setMouseCallback(winname, on_mouse, userdata);
 }
 
 // =============================================================================
@@ -513,220 +502,5 @@ RotatedRect cv_camshift(cv::Mat* bp_image, Rect crect, cv::TermCriteria* criteri
 // =============================================================================
 //   MSER
 // =============================================================================
-
-void* cv_mser_new(int delta,
-                  int min_area,
-                  int max_area,
-                  double max_variation,
-                  double min_diversity,
-                  int max_evolution,
-                  double area_threshold,
-                  double min_margin,
-                  int edge_blur_size) {
-    cv::Ptr<cv::MSER> result = cv::MSER::create(delta,
-                                                min_area,
-                                                max_area,
-                                                max_variation,
-                                                min_diversity,
-                                                max_evolution,
-                                                area_threshold,
-                                                min_margin,
-                                                edge_blur_size);
-    return new cv::Ptr<cv::MSER>(result);
-}
-
-void cv_mser_drop(cv::Ptr<cv::MSER>* detector) {
-    delete detector;
-    detector = nullptr;
-}
-
-void cv_mser_detect_regions(cv::Ptr<cv::MSER>* detector,
-                            cv::Mat* image,
-                            CVec<CVec<Point2i>>* msers,
-                            CVec<Rect>* bboxes) {
-    std::vector<std::vector<cv::Point>> msers_vector;
-    std::vector<cv::Rect> bboxes_vector;
-
-    detector->get()->detectRegions(*image, msers_vector, bboxes_vector);
-
-    cv_to_ffi(msers_vector, msers);
-    cv_to_ffi(bboxes_vector, bboxes);
-}
-
-void cv_mser_detect_and_compute(cv::Ptr<cv::MSER>* detector,
-                                cv::Mat* image,
-                                cv::Mat* mask,
-                                CVec<KeyPoint>* keypoints,
-                                cv::Mat* descriptors,
-                                bool useProvidedKeypoints) {
-    std::vector<cv::KeyPoint> keypoints_vector;
-    detector->get()->detectAndCompute(*image, *mask, keypoints_vector, *descriptors, useProvidedKeypoints);
-    cv_to_ffi(keypoints_vector, keypoints);
-}
-
-void* cv_surf_new(double hessianThreshold, int nOctaves, int nOctaveLayers, bool extended, bool upright) {
-    auto result = cv::xfeatures2d::SURF::create(hessianThreshold, nOctaves, nOctaveLayers, extended, upright);
-    return new cv::Ptr<cv::xfeatures2d::SURF>(result);
-}
-void cv_surf_drop(cv::Ptr<cv::xfeatures2d::SURF>* detector) {
-    delete detector;
-    detector = nullptr;
-}
-
-void cv_surf_detect_and_compute(cv::Ptr<cv::xfeatures2d::SURF>* detector,
-                                cv::Mat* image,
-                                cv::Mat* mask,
-                                CVec<KeyPoint>* keypoints,
-                                cv::Mat* descriptors,
-                                bool useProvidedKeypoints) {
-    std::vector<cv::KeyPoint> keypoints_vector;
-    detector->get()->detectAndCompute(*image, *mask, keypoints_vector, *descriptors, useProvidedKeypoints);
-    cv_to_ffi(keypoints_vector, keypoints);
-}
-
-void* cv_sift_new(int nfeatures, int nOctaveLayers, double contrastThreshold, double edgeThreshold, double sigma) {
-    auto result = cv::xfeatures2d::SIFT::create(nfeatures, nOctaveLayers, contrastThreshold, edgeThreshold, sigma);
-    return new cv::Ptr<cv::xfeatures2d::SIFT>(result);
-}
-void cv_sift_drop(cv::Ptr<cv::xfeatures2d::SIFT>* detector) {
-    delete detector;
-    detector = nullptr;
-}
-
-void cv_sift_detect_and_compute(cv::Ptr<cv::xfeatures2d::SIFT>* detector,
-                                cv::Mat* image,
-                                cv::Mat* mask,
-                                CVec<KeyPoint>* keypoints,
-                                cv::Mat* descriptors,
-                                bool useProvidedKeypoints) {
-    std::vector<cv::KeyPoint> keypoints_vector;
-    detector->get()->detectAndCompute(*image, *mask, keypoints_vector, *descriptors, useProvidedKeypoints);
-    cv_to_ffi(keypoints_vector, keypoints);
-}
-
-void* cv_matcher_new(const char* descriptorMatcherType) {
-    auto result = cv::DescriptorMatcher::create(descriptorMatcherType);
-    return new cv::Ptr<cv::DescriptorMatcher>(result);
-}
-
-void cv_matcher_drop(cv::Ptr<cv::DescriptorMatcher>* descriptorMatcher) {
-    delete descriptorMatcher;
-    descriptorMatcher = nullptr;
-}
-
-void cv_matcher_add(cv::Ptr<cv::DescriptorMatcher>& descriptorMatcher, CVec<cv::Mat*>& descriptors) {
-    std::vector<cv::Mat> descriptors_vector;
-    ffi_to_cv(descriptors, &descriptors_vector);
-    descriptorMatcher.get()->add(descriptors_vector);
-}
-
-void cv_matcher_train(cv::Ptr<cv::DescriptorMatcher>& descriptorMatcher) {
-    descriptorMatcher.get()->train();
-}
-
-bool cv_matcher_is_empty(cv::Ptr<cv::DescriptorMatcher>& descriptorMatcher) {
-    return descriptorMatcher.get()->empty();
-}
-
-void cv_matcher_match(cv::Ptr<cv::DescriptorMatcher>& descriptorMatcher,
-                      cv::Mat& queryDescriptors,
-                      CVec<DMatch>* matches) {
-    std::vector<cv::DMatch> matches_vector;
-    descriptorMatcher.get()->match(queryDescriptors, matches_vector);
-    cv_to_ffi(matches_vector, matches);
-}
-
-void cv_matcher_match_two(cv::Ptr<cv::DescriptorMatcher>& descriptorMatcher,
-                          cv::Mat& queryDescriptors,
-                          cv::Mat& trainDescriptors,
-                          CVec<DMatch>* matches) {
-    std::vector<cv::DMatch> matches_vector;
-    descriptorMatcher.get()->match(queryDescriptors, trainDescriptors, matches_vector);
-    cv_to_ffi(matches_vector, matches);
-}
-
-void cv_ocr_run(cv::Ptr<cv::text::BaseOCR>& ocr,
-                cv::Mat& image,
-                CDisposableString* output_text,
-                CVec<Rect>* component_rects,
-                CVec<CDisposableString>* component_texts,
-                CVec<float>* component_confidences,
-                int component_level) {
-    std::string output;
-    std::vector<cv::Rect> boxes;
-    std::vector<std::string> words;
-    std::vector<float> confidences;
-    ocr.get()->run(image, output, &boxes, &words, &confidences, component_level);
-
-    cv_to_ffi(output, output_text);
-    cv_to_ffi(boxes, component_rects);
-    cv_to_ffi(words, component_texts);
-    cv_to_ffi(confidences, component_confidences);
-}
-
-void cv_matcher_knn_match(cv::Ptr<cv::DescriptorMatcher>& descriptorMatcher,
-                          cv::Mat& queryDescriptors,
-                          int k,
-                          CVec<CVec<DMatch>>* matches) {
-    std::vector<std::vector<cv::DMatch>> matches_vector;
-    descriptorMatcher.get()->knnMatch(queryDescriptors, matches_vector, k);
-    cv_to_ffi(matches_vector, matches);
-}
-
-void cv_tesseract_new(const char* datapath,
-                      const char* language,
-                      const char* char_whitelist,
-                      int oem,
-                      int psmode,
-                      Result<void*>* result) {
-    *result = Result<void*>::FromFunction([datapath, language, char_whitelist, oem, psmode]() {
-        auto result = cv::text::OCRTesseract::create(datapath, language, char_whitelist, oem, psmode);
-        return new cv::Ptr<cv::text::OCRTesseract>(result);
-    });
-}
-
-void cv_tesseract_drop(cv::Ptr<cv::text::OCRTesseract>* ocr) {
-    delete ocr;
-    ocr = nullptr;
-}
-
-void cv_hmm_new(const char* classifier_filename,
-                const char* vocabulary,
-                cv::Mat& transition_probabilities_table,
-                cv::Mat& emission_probabilities_table,
-                cv::text::classifier_type classifier_type,
-                Result<void*>* result) {
-    *result = Result<void*>::FromFunction([classifier_filename,
-                                           vocabulary,
-                                           transition_probabilities_table,
-                                           emission_probabilities_table,
-                                           classifier_type]() {
-        std::string voc(vocabulary);
-        auto classifier = cv::text::loadOCRHMMClassifier(classifier_filename, classifier_type);
-        auto result = cv::text::OCRHMMDecoder::create(
-            classifier, voc, transition_probabilities_table, emission_probabilities_table);
-        return new cv::Ptr<cv::text::OCRHMMDecoder>(result);
-    });
-}
-
-void cv_hmm_drop(cv::Ptr<cv::text::OCRHMMDecoder>* ocr) {
-    delete ocr;
-    ocr = nullptr;
-}
-
-void cv_holistic_new(const char* archive_file,
-                     const char* weights_file,
-                     const char* words_file,
-                     Result<void*>* result) {
-    *result = Result<void*>::FromFunction([archive_file, weights_file, words_file]() {
-        auto result = cv::text::OCRHolisticWordRecognizer::create(archive_file, weights_file, words_file);
-        return new cv::Ptr<cv::text::OCRHolisticWordRecognizer>(result);
-    });
-}
-
-void cv_holistic_drop(cv::Ptr<cv::text::OCRHolisticWordRecognizer>* ocr) {
-    delete ocr;
-    ocr = nullptr;
-}
 
 EXTERN_C_END
